backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pulse-news-backend"
version = "0.1.0"
description = "Backend for Pulse News"
requires-python = ">=3.12"
dependencies = [
  "apscheduler>=3.11.2,<4.0.0",
  "fastapi>=0.135.3,<1.0.0",
  "uvicorn[standard]>=0.34.2,<1.0.0",
  "sqlalchemy>=2.0.49,<3.0.0",
  "pydantic-settings>=2.11.0,<3.0.0",
  "itsdangerous>=2.2.0,<3.0.0",
  "litellm>=1.83.4,<2.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0.0",
  "pytest>=8.3.5,<9.0.0",
  "ruff>=0.11.8,<1.0.0",
]

[tool.setuptools]
packages = ["app"]
package-dir = {"" = "."}

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
